Move utilities to lib
