+++ package.json
@@ -5,6 +5,7 @@
   "type": "module",
   "scripts": {},
   "dependencies": {
-    "treemap-squarify": "^1.0.1"
+    "treemap-squarify": "^1.0.1",
+    "xlsx": "^0.18.5"
   }
 }
