+++ public/scripts/lib/ColorTool.js
