+++ public/scripts/lib/ObjectSet.js
