+++ public/scripts/lib/SVGUtil.js
